requirements.txt
pandas
numpy
matplotlib
pyarrow
PyYAML
tqdm
seaborn